Refactor txpool tests Signed-off-by: Fabio Di Fabio <fabio.difabio@consensys.net>
