+++ ethereum/eth/src/test/java/org/hyperledger/besu/ethereum/eth/transactions/TransactionPoolLondonTest.java
@@ -1,0 +1,205 @@
+/*
+ * Copyright Hyperledger Besu Contributors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+package org.hyperledger.besu.ethereum.eth.transactions;
+
+import static java.util.Arrays.asList;
+import static java.util.Collections.emptyList;
+import static java.util.stream.Collectors.toList;
+import static org.mockito.Mockito.when;
+
+import org.hyperledger.besu.config.StubGenesisConfigOptions;
+import org.hyperledger.besu.datatypes.Wei;
+import org.hyperledger.besu.ethereum.core.Block;
+import org.hyperledger.besu.ethereum.core.BlockBody;
+import org.hyperledger.besu.ethereum.core.BlockHeader;
+import org.hyperledger.besu.ethereum.core.BlockHeaderBuilder;
+import org.hyperledger.besu.ethereum.core.BlockHeaderTestFixture;
+import org.hyperledger.besu.ethereum.core.Difficulty;
+import org.hyperledger.besu.ethereum.core.ExecutionContextTestFixture;
+import org.hyperledger.besu.ethereum.core.PrivacyParameters;
+import org.hyperledger.besu.ethereum.core.Transaction;
+import org.hyperledger.besu.ethereum.core.TransactionReceipt;
+import org.hyperledger.besu.ethereum.core.TransactionTestFixture;
+import org.hyperledger.besu.ethereum.eth.transactions.sorter.AbstractPendingTransactionsSorter;
+import org.hyperledger.besu.ethereum.eth.transactions.sorter.BaseFeePendingTransactionsSorter;
+import org.hyperledger.besu.ethereum.mainnet.MainnetBlockHeaderFunctions;
+import org.hyperledger.besu.ethereum.mainnet.ProtocolSchedule;
+import org.hyperledger.besu.ethereum.mainnet.ProtocolScheduleBuilder;
+import org.hyperledger.besu.ethereum.mainnet.ProtocolSpecAdapters;
+import org.hyperledger.besu.ethereum.mainnet.feemarket.FeeMarket;
+import org.hyperledger.besu.evm.internal.EvmConfiguration;
+import org.hyperledger.besu.plugin.data.TransactionType;
+import org.hyperledger.besu.testutil.TestClock;
+
+import java.math.BigInteger;
+import java.time.ZoneId;
+import java.util.List;
+import java.util.Optional;
+import java.util.function.Function;
+
+import org.junit.Ignore;
+import org.junit.Test;
+
+public class TransactionPoolLondonTest extends AbstractTransactionPoolTest {
+
+  @Override
+  protected AbstractPendingTransactionsSorter createPendingTransactionsSorter() {
+
+    return new BaseFeePendingTransactionsSorter(
+        ImmutableTransactionPoolConfiguration.builder()
+            .txPoolMaxSize(MAX_TRANSACTIONS)
+            .txPoolLimitByAccountPercentage(1)
+            .build(),
+        TestClock.system(ZoneId.systemDefault()),
+        metricsSystem,
+        protocolContext.getBlockchain()::getChainHeadHeader);
+  }
+
+  @Override
+  protected Transaction createTransaction(
+      final int transactionNumber, final Optional<BigInteger> maybeChainId) {
+    return createBaseTransaction(transactionNumber)
+        .chainId(maybeChainId)
+        .createTransaction(KEY_PAIR1);
+  }
+
+  @Override
+  protected Transaction createTransaction(final int transactionNumber, final Wei maxPrice) {
+    return createBaseTransaction(transactionNumber)
+        .maxFeePerGas(Optional.of(maxPrice))
+        .maxPriorityFeePerGas(Optional.of(maxPrice.divide(5L)))
+        .createTransaction(KEY_PAIR1);
+  }
+
+  @Override
+  protected TransactionTestFixture createBaseTransaction(final int transactionNumber) {
+    return new TransactionTestFixture()
+        .nonce(transactionNumber)
+        .gasLimit(blockGasLimit)
+        .gasPrice(null)
+        .maxFeePerGas(Optional.of(Wei.of(5000L)))
+        .maxPriorityFeePerGas(Optional.of(Wei.of(1000L)))
+        .type(TransactionType.EIP1559);
+  }
+
+  @Override
+  protected ExecutionContextTestFixture createExecutionContextTestFixture() {
+    final ProtocolSchedule protocolSchedule =
+        new ProtocolScheduleBuilder(
+                new StubGenesisConfigOptions().londonBlock(0L).baseFeePerGas(10L),
+                BigInteger.valueOf(1),
+                ProtocolSpecAdapters.create(0, Function.identity()),
+                new PrivacyParameters(),
+                false,
+                false,
+                EvmConfiguration.DEFAULT)
+            .createProtocolSchedule();
+    final ExecutionContextTestFixture executionContextTestFixture =
+        ExecutionContextTestFixture.builder().protocolSchedule(protocolSchedule).build();
+
+    final Block block =
+        new Block(
+            new BlockHeaderTestFixture()
+                .gasLimit(
+                    executionContextTestFixture
+                        .getBlockchain()
+                        .getChainHeadBlock()
+                        .getHeader()
+                        .getGasLimit())
+                .difficulty(Difficulty.ONE)
+                .baseFeePerGas(Wei.of(10L))
+                .parentHash(executionContextTestFixture.getBlockchain().getChainHeadHash())
+                .number(executionContextTestFixture.getBlockchain().getChainHeadBlockNumber() + 1)
+                .buildHeader(),
+            new BlockBody(List.of(), List.of()));
+    executionContextTestFixture.getBlockchain().appendBlock(block, List.of());
+
+    return executionContextTestFixture;
+  }
+
+  @Override
+  protected FeeMarket getFeeMarket() {
+    return FeeMarket.london(0L, Optional.of(Wei.of(0L)));
+  }
+
+  @Override
+  protected Block appendBlock(
+      final Difficulty difficulty,
+      final BlockHeader parentBlock,
+      final Transaction... transactionsToAdd) {
+    final List<Transaction> transactionList = asList(transactionsToAdd);
+    final Block block =
+        new Block(
+            new BlockHeaderTestFixture()
+                .baseFeePerGas(Wei.of(10L))
+                .gasLimit(parentBlock.getGasLimit())
+                .difficulty(difficulty)
+                .parentHash(parentBlock.getHash())
+                .number(parentBlock.getNumber() + 1)
+                .buildHeader(),
+            new BlockBody(transactionList, emptyList()));
+    final List<TransactionReceipt> transactionReceipts =
+        transactionList.stream()
+            .map(transaction -> new TransactionReceipt(1, 1, emptyList(), Optional.empty()))
+            .collect(toList());
+    blockchain.appendBlock(block, transactionReceipts);
+    return block;
+  }
+
+  @Test
+  public void shouldAcceptZeroGasPriceFrontierTxsWhenMinGasPriceIsZeroAndLondonWithZeroBaseFee() {
+    when(miningParameters.getMinTransactionGasPrice()).thenReturn(Wei.ZERO);
+    when(protocolSpec.getFeeMarket()).thenReturn(FeeMarket.london(0, Optional.of(Wei.ZERO)));
+    whenBlockBaseFeeIsZero();
+
+    final Transaction transaction =
+        createBaseTransaction(0)
+            .gasPrice(Wei.ZERO)
+            .type(TransactionType.FRONTIER)
+            .createTransaction(KEY_PAIR1);
+
+    givenTransactionIsValid(transaction);
+    assertLocalTransactionValid(transaction);
+  }
+
+  @Test
+  public void shouldAcceptZeroGasPrice1559TxsWhenMinGasPriceIsZeroAndLondonWithZeroBaseFee() {
+    when(miningParameters.getMinTransactionGasPrice()).thenReturn(Wei.ZERO);
+    when(protocolSpec.getFeeMarket()).thenReturn(FeeMarket.london(0, Optional.of(Wei.ZERO)));
+    whenBlockBaseFeeIsZero();
+
+    final Transaction transaction = createTransaction(0, Wei.ZERO);
+
+    givenTransactionIsValid(transaction);
+    assertLocalTransactionValid(transaction);
+  }
+
+  @Test
+  @Override
+  @Ignore
+  public void shouldRejectLocalTransactionIfFeeCapExceeded() {
+    // ignore since this is going to fail until the branch with the fix is released
+  }
+
+  private void whenBlockBaseFeeIsZero() {
+    final BlockHeader header =
+        BlockHeaderBuilder.fromHeader(blockchain.getChainHeadHeader())
+            .baseFee(Wei.ZERO)
+            .blockHeaderFunctions(new MainnetBlockHeaderFunctions())
+            .parentHash(blockchain.getChainHeadHash())
+            .buildBlockHeader();
+    blockchain.appendBlock(new Block(header, BlockBody.empty()), emptyList());
+  }
+}
